+++ crates/teloxide-core/src/adaptors/erased.rs
@@ -232,6 +232,7 @@ where
         restrict_chat_member,
         promote_chat_member,
         set_chat_administrator_custom_title,
+        set_chat_member_tag,
         ban_chat_sender_chat,
         unban_chat_sender_chat,
         set_chat_permissions,
@@ -609,6 +610,12 @@ trait ErasableRequester<'a> {
         chat_id: Recipient,
         user_id: UserId,
     ) -> ErasedRequest<'a, PromoteChatMember, Self::Err>;
+
+    fn set_chat_member_tag(
+        &self,
+        chat_id: Recipient,
+        user_id: UserId,
+    ) -> ErasedRequest<'a, SetChatMemberTag, Self::Err>;
 
     fn set_chat_administrator_custom_title(
         &self,
@@ -1641,6 +1648,14 @@ where
         user_id: UserId,
     ) -> ErasedRequest<'a, PromoteChatMember, Self::Err> {
         Requester::promote_chat_member(self, chat_id, user_id).erase()
+    }
+
+    fn set_chat_member_tag(
+        &self,
+        chat_id: Recipient,
+        user_id: UserId,
+    ) -> ErasedRequest<'a, SetChatMemberTag, Self::Err> {
+        Requester::set_chat_member_tag(self, chat_id, user_id).erase()
     }
 
     fn set_chat_administrator_custom_title(