+++ crates/teloxide-core/src/types/update.rs
@@ -633,6 +633,7 @@ mod test {
                     }),
                     reply_markup: None,
                     author_signature: None,
+                    sender_tag: None,
                     paid_star_count: None,
                     effect_id: None,
                     is_automatic_forward: false,