+++ crates/teloxide/src/dispatching/handler_ext.rs
@@ -227,6 +227,7 @@ mod tests {
                     }),
                     reply_markup: None,
                     author_signature: None,
+                    sender_tag: None,
                     paid_star_count: None,
                     effect_id: None,
                     is_automatic_forward: false,